services/svc-forecast/requirements.txt
fastapi
uvicorn
psycopg[binary]
sqlalchemy
pydantic-settings
passlib[bcrypt]
email-validator
